requirements.txt
# from github.com/jimbobbenett

azure-cognitiveservices-speech
python-dotenv